NgoLeHoangminiclouddemo/application-backend-server/requirements.txt
Flask==2.3.3
requests==2.31.0
python-jose==3.3.0
mysql-connector-python==8.1.0
